pyproject.toml
[tool.poetry]
name = "test_task"
version = "0.1.0"
description = ""
authors = ["Nikolay Nikishev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
sqlalchemy = "1.4.41"
alembic = "^1.12.0"
asyncpg = "^0.30.0"
greenlet = "^3.2.3"
fastapi-utils = "^0.8.0"
typing-inspect = "^0.9.0"
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
